+++ packages/core/src/schemas/index.ts
@@ -1,4 +1,5 @@
 export * from "./issues";
+export * from "./notifications";
 export * from "./workspace";
 export * from "./planning";
 export * from "./activity";